WebDavMountsScreen: M3, refactoring
