+++ app/src/main/kotlin/at/bitfire/davdroid/db/WebDavMountDao.kt
@@ -4,11 +4,11 @@
 
 package at.bitfire.davdroid.db
 
-import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
+import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface WebDavMountDao {
@@ -20,12 +20,18 @@ interface WebDavMountDao {
     fun getAll(): List<WebDavMount>
 
     @Query("SELECT * FROM webdav_mount ORDER BY name, url")
-    fun getAllLive(): LiveData<List<WebDavMount>>
+    fun getAllFlow(): Flow<List<WebDavMount>>
 
     @Query("SELECT * FROM webdav_mount WHERE id=:id")
     fun getById(id: Long): WebDavMount
 
     @Insert
     fun insert(mount: WebDavMount): Long
+
+
+    // complex queries
+
+    @Query("SELECT * FROM webdav_mount ORDER BY name, url")
+    fun getAllWithRootDocumentFlow(): Flow<List<WebDavMountWithRootDocument>>
 
 }
